[TASK] Extract garbage removall from GarbageCollector
This pr:

* Extract's the garbage removall from the GarbageCollector

Fixes: #2114
